requirements.txt
asgiref==3.5.1
asttokens==2.0.5
async-timeout==4.0.2
attrs==21.4.0
autobahn==22.5.1
Automat==20.2.0
backcall==0.2.0
boto==2.49.0
boto3==1.24.46
botocore==1.27.46
Brotli==1.0.9
cairocffi==1.3.0
CairoSVG==2.5.2
certifi==2022.5.18
cffi==1.15.0
channels
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.5
constantly==15.1.0
cryptography==37.0.2
cssselect==1.1.0
cssselect2==0.6.0
daphne==3.0.2
dash==2.5.1
dash-bootstrap-components==0.13.1
dash-core-components==2.0.0
dash-daq==0.5.0
dash-html-components==2.0.0
dash-renderer==1.9.1
dash-table==5.0.0
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.13
Django==3.2.10
django-allauth==0.50.0
django-bower==5.2.0
django-decouple==2.1
django-des==2.4.1
django-invitations==1.9.3
django-jsignature==0.11
django-nvd3==0.9.7
django-plotly-dash==2.0.0
django-redis==5.2.0
django-solo==1.1.3
django-storages==1.13.1
django-widget-tweaks==1.4.12
djongo==1.3.6
dnspython==2.2.1
dpd-components==0.1.0
executing==0.8.3
Flask==2.1.2
Flask-Compress==1.12
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
ipython==8.4.0
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
jmespath==1.0.1
lxml==4.8.0
MarkupSafe==2.1.1
matplotlib-inline==0.1.3
msgpack==1.0.4
numpy==1.22.4
oauthlib==3.2.0
packaging==21.3
pandas==1.4.2
parso==0.8.3
pickleshare==0.7.5
Pillow==9.1.0
plotly==5.8.2
prompt-toolkit==3.0.29
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
Pygments==2.12.0
PyJWT==2.4.0
pymongo==4.1.1
pyOpenSSL==22.0.0
pyparsing==3.0.9
pyquery==1.4.3
python-dateutil==2.8.2
python-nvd3==0.14.2
python-slugify==1.1.4
python3-openid==3.2.0
pytz==2022.1
redis==4.3.3
requests==2.27.1
requests-oauthlib==1.3.1
s3transfer==0.6.0
service-identity==21.1.0
six==1.16.0
sqlparse==0.2.4
stack-data==0.3.0
svgutils==0.3.4
sweetify==2.3.1
tenacity==8.0.1
tinycss2==1.1.1
traitlets==5.3.0
Twisted==22.4.0
twisted-iocpsupport==1.0.2
txaio==22.2.1
typing_extensions==4.2.0
tzdata==2022.1
Unidecode==1.3.4
urllib3==1.26.9
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.3
wrapt==1.14.1
zope.interface==5.4.0
